tests/requirements.txt
# ============================================
# Python Testing Dependencies
# ============================================

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-mock==3.12.0
pytest-env==1.1.3
pytest-json-report==1.5.0

# HTTP testing
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Docker testing
docker==6.1.3
testcontainers==3.7.1

# Database testing
# Note: Requires Python 3.10-3.13 (Python 3.14 not yet supported by ecosystem)
psycopg2-binary==2.9.9  # PostgreSQL adapter
sqlalchemy==2.0.23

# Validation
pydantic==2.5.2
jsonschema==4.20.0
pyyaml==6.0.1

# Keycloak/OAuth testing
python-keycloak==3.8.0
pyjwt==2.8.0

# Prometheus metrics testing
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0

# Mocking
responses==0.24.1
faker==20.1.0

# Performance profiling
memory-profiler==0.61.0

